Moved controls to ui components folder
