--- src/components/Controls/Form/Form.scss
@@ -1,4 +1,0 @@
-.form {
-  display: flex;
-  align-items: flex-end;
-}
