--- src/components/Controls/FormControls.variables.scss
@@ -1,8 +1,0 @@
-$form-control-padding: 8px;
-
-$form-control-border-width: 1px;
-$form-control-border-radius: 5px;
-
-$form-control-transition: 0.5s;
-
-$form-control-box-shadow: 0 0 5px;
