+++ src/components/Controls/FormGroup/FormGroup.scss
@@ -1,5 +1,3 @@
-@import '../FormControls.variables.scss';
-
 .form-group {
   display: flex;
   flex-direction: column;
