--- src/components/Controls/Input/Input.scss
@@ -1,34 +1,0 @@
-@import '../../../variables.scss';
-@import '../FormControls.variables.scss';
-
-.input {
-  width: 100%;
-  padding: $form-control-padding;
-  border: $form-control-border-width solid $color-middle-green-50;
-  border-radius: $form-control-border-radius;
-  transition: $form-control-transition;
-  outline: none;
-  background-color: transparent;
-
-  &_small {
-    height: 25px;
-  }
-
-  &_normal {
-    height: 35px;
-  }
-
-  &:disabled {
-    border: $form-control-border-width solid $color-middle-green-25;
-    color: $color-middle-green-25;
-  }
-
-  &:focus {
-    border-color: $color-middle-green;
-    box-shadow: $form-control-box-shadow $color-middle-green;
-  }
-
-  &::placeholder {
-    color: $color-middle-green-50;
-  }
-}
