--- src/components/Controls/Label/Label.scss
@@ -1,5 +1,0 @@
-@import '../FormControls.variables.scss';
-
-.label {
-  padding: $form-control-padding;
-}
