+++ src/components/Controls/index.ts
@@ -1,25 +1,3 @@
-import Form from './Form';
 import FormGroup from './FormGroup';
-import Label from './Label';
-import Button from './Button';
-import Input from './Input';
-import { DropdownList, DropdownMenu, productDropdownItemRenderer, categoryDropdownItemRenderer } from './Dropdown';
-import DropdownItem from './DropdownItem';
-import Checkbox from './Checkbox/input';
-import { Table, TableColumn } from './Table';
 
-export {
-  Form,
-  FormGroup,
-  Label,
-  Button,
-  Input,
-  DropdownList,
-  DropdownMenu,
-  DropdownItem,
-  Checkbox,
-  Table,
-  TableColumn,
-};
-
-export { productDropdownItemRenderer, categoryDropdownItemRenderer };
+export { FormGroup };
