+++ src/components/NoteInput/NoteInput.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import './NoteInput.scss';
-import { FormGroup, Label, Input, Button, productDropdownItemRenderer, DropdownList } from '../Controls';
 import { NoteInputStateToPropsMapResult, NoteInputDispatchToPropsMapResult } from './NoteInputConnected';
 import { MealType } from '../../models';
 import { NotesOperationsActionTypes } from '../../action-types';
 import { useDebounce, useNoteValidation, useIdFromRoute } from '../../hooks';
-import { Loader } from '../__ui__';
+import { Loader, Label, Input, Button, productDropdownItemRenderer, DropdownList } from '../__ui__';
+import { FormGroup } from '../Controls';
 
 interface NoteInputProps extends NoteInputStateToPropsMapResult, NoteInputDispatchToPropsMapResult {
   mealType: MealType;
