+++ src/components/__ui__/Checkbox/Checkbox.scss
@@ -1,5 +1,5 @@
 @import '../../../variables.scss';
-@import '../FormControls.variables.scss';
+@import '../controls.variables.scss';
 
 .checkbox-container {
   height: 20px;
@@ -49,7 +49,7 @@
   }
 
   & &__input ~ &__checkmark {
-    transition: $form-control-transition;
+    transition: $control-transition;
   }
 
   // Checkbox is hovered
