+++ src/components/__ui__/Label/Label.scss
@@ -1,0 +1,5 @@
+@import '../controls.variables.scss';
+
+.label {
+  padding: $control-padding;
+}
