+++ src/components/__ui__/controls.variables.scss
@@ -1,0 +1,5 @@
+$control-padding: 8px;
+$control-border-width: 1px;
+$control-border-radius: 5px;
+$control-transition: 0.5s;
+$control-box-shadow: 0 0 5px;
