+++ src/components/__ui__/index.ts
@@ -3,5 +3,29 @@ import Badge from './Badge';
 import Icon from './Icon';
 import Loader from './Loader';
 import Pagination from './Pagination';
+import Button from './Button';
+import Checkbox from './Checkbox';
+import { DropdownList, DropdownMenu, categoryDropdownItemRenderer, productDropdownItemRenderer } from './Dropdown';
+import DropdownItem from './DropdownItem';
+import Input from './Input';
+import Label from './Label';
+import { Table, TableColumn } from './Table';
 
-export { Container, Badge, Icon, Loader, Pagination };
+export {
+  Container,
+  Badge,
+  Icon,
+  Loader,
+  Pagination,
+  Button,
+  Checkbox,
+  DropdownList,
+  DropdownMenu,
+  DropdownItem,
+  Input,
+  Label,
+  Table,
+  TableColumn,
+};
+
+export { categoryDropdownItemRenderer, productDropdownItemRenderer };
